+++ code/classeine-lib/clsn/ui/renderers/ToggleButtonRenderer.h
@@ -1,27 +1,20 @@
 #pragma once
 
-#include "BorderRenderer.h"
 #include "ButtonRenderer.h"
 
-#include "clsn/ui/Graphics.h"
-#include "clsn/ui/ToggleButton.h"
-
-#include "clsn/draw/Region.h"
-
-#include "clsn/ui/IRenderer.h"
+namespace  clsn::ui
+{
+    class ClickableControl;
+}
 
 namespace clsn::ui::renderers
 {
     using namespace clsn::draw;
     using namespace clsn::ui;
 
-    template <typename LabelRendererType>
-    class ToggleButtonRenderer : public ButtonRenderer<LabelRendererType>
+    class ToggleButtonRenderer : public ButtonRenderer
     {
     public:
-        [[nodiscard]] virtual auto paintAsPressed(const ClickableControl& button) const -> bool
-        {
-            return static_cast<const ToggleButton&>(button).isChecked() || button.isPressed();
-        }
+        [[nodiscard]] virtual auto paintAsPressed(const ClickableControl& button) const -> bool;
     };
 }